Extract PreferencesObserver and ScopesObserver to RunnerStore+Observers.swift
Moves the two private @MainActor observer classes out of RunnerStore.swift
into a dedicated RunnerStore+Observers.swift, following the existing +PollLoop /
+PollBridge / +InstallPathMap file-per-concern convention.

Motivation: RunnerStore.swift was 624 lines, 4 over the SwiftLint file_length
limit of 620. Removing doc-comments would have been the wrong fix — the comments
explain non-obvious Swift 6 concurrency constraints. A file split is the correct
response and leaves RunnerStore.swift at ~562 lines.

No logic changes. PreferencesObserver and ScopesObserver remain private to the
RunnerBar module; Swift's file-private vs module-private distinction means the
classes are still accessible from RunnerStore.swift within the same module.
