+++ Sources/RunnerBar/Runner/RunnerStore+Observers.swift
@@ -1,0 +1,73 @@
+// RunnerStore+Observers.swift
+// RunnerBar
+import Foundation
+
+// MARK: - PreferencesObserver
+
+/// Drives a recursive `withObservationTracking` loop for `AppPreferencesStoreProtocol.pollingInterval`
+/// entirely on the `@MainActor`. Because every method is `@MainActor`-isolated, the local
+/// `func observe()` inside `start()` is implicitly `@MainActor` — no `@Sendable` annotation
+/// is required and no value crosses an isolation boundary.
+@MainActor
+final class PreferencesObserver {
+    /// The continuation used to push new `pollingInterval` values into the `AsyncStream`.
+    private let continuation: AsyncStream<Int>.Continuation
+    /// The injected preferences store — avoids singleton access inside the observer.
+    private let store: any AppPreferencesStoreProtocol
+
+    /// Creates a new observer that writes changes into `continuation`.
+    init(continuation: AsyncStream<Int>.Continuation, store: any AppPreferencesStoreProtocol) {
+        self.continuation = continuation
+        self.store = store
+    }
+
+    /// Registers a single `withObservationTracking` pass and re-registers itself on change.
+    func start() {
+        func observe() {
+            withObservationTracking {
+                _ = store.pollingInterval
+            } onChange: { [weak self] in
+                Task { @MainActor [weak self] in
+                    guard let self else { return }
+                    self.continuation.yield(self.store.pollingInterval)
+                    self.start()
+                }
+            }
+        }
+        observe()
+    }
+}
+
+// MARK: - ScopesObserver
+
+/// Drives a recursive `withObservationTracking` loop for `ScopeStoreProtocol.activeScopes`
+/// entirely on the `@MainActor`. Same isolation rationale as `PreferencesObserver`.
+@MainActor
+final class ScopesObserver {
+    /// The continuation used to push new `activeScopes` values into the `AsyncStream`.
+    private let continuation: AsyncStream<[String]>.Continuation
+    /// The injected scope store — avoids singleton access inside the observer.
+    private let store: any ScopeStoreProtocol
+
+    /// Creates a new observer that writes changes into `continuation`.
+    init(continuation: AsyncStream<[String]>.Continuation, store: any ScopeStoreProtocol) {
+        self.continuation = continuation
+        self.store = store
+    }
+
+    /// Registers a single `withObservationTracking` pass and re-registers itself on change.
+    func start() {
+        func observe() {
+            withObservationTracking {
+                _ = store.activeScopes
+            } onChange: { [weak self] in
+                Task { @MainActor [weak self] in
+                    guard let self else { return }
+                    self.continuation.yield(self.store.activeScopes)
+                    self.start()
+                }
+            }
+        }
+        observe()
+    }
+}
