+++ Sources/RunnerBar/Runner/RunnerStore.swift
@@ -65,74 +65,6 @@ protocol ScopeStoreProtocol: AnyObject, Sendable {
 /// injected at the production call site without any wrapper.
 extension ScopeStore: ScopeStoreProtocol {}
 
-// MARK: - Observation helpers
-
-/// Drives a recursive `withObservationTracking` loop for `AppPreferencesStoreProtocol.pollingInterval`
-/// entirely on the `@MainActor`. Because every method is `@MainActor`-isolated, the local
-/// `func observe()` inside `start()` is implicitly `@MainActor` — no `@Sendable` annotation
-/// is required and no value crosses an isolation boundary.
-@MainActor
-private final class PreferencesObserver {
-    /// The continuation used to push new `pollingInterval` values into the `AsyncStream`.
-    private let continuation: AsyncStream<Int>.Continuation
-    /// The injected preferences store — avoids singleton access inside the observer.
-    private let store: any AppPreferencesStoreProtocol
-
-    /// Creates a new observer that writes changes into `continuation`.
-    init(continuation: AsyncStream<Int>.Continuation, store: any AppPreferencesStoreProtocol) {
-        self.continuation = continuation
-        self.store = store
-    }
-
-    /// Registers a single `withObservationTracking` pass and re-registers itself on change.
-    func start() {
-        func observe() {
-            withObservationTracking {
-                _ = store.pollingInterval
-            } onChange: { [weak self] in
-                Task { @MainActor [weak self] in
-                    guard let self else { return }
-                    self.continuation.yield(self.store.pollingInterval)
-                    self.start()
-                }
-            }
-        }
-        observe()
-    }
-}
-
-/// Drives a recursive `withObservationTracking` loop for `ScopeStoreProtocol.activeScopes`
-/// entirely on the `@MainActor`. Same isolation rationale as `PreferencesObserver`.
-@MainActor
-private final class ScopesObserver {
-    /// The continuation used to push new `activeScopes` values into the `AsyncStream`.
-    private let continuation: AsyncStream<[String]>.Continuation
-    /// The injected scope store — avoids singleton access inside the observer.
-    private let store: any ScopeStoreProtocol
-
-    /// Creates a new observer that writes changes into `continuation`.
-    init(continuation: AsyncStream<[String]>.Continuation, store: any ScopeStoreProtocol) {
-        self.continuation = continuation
-        self.store = store
-    }
-
-    /// Registers a single `withObservationTracking` pass and re-registers itself on change.
-    func start() {
-        func observe() {
-            withObservationTracking {
-                _ = store.activeScopes
-            } onChange: { [weak self] in
-                Task { @MainActor [weak self] in
-                    guard let self else { return }
-                    self.continuation.yield(self.store.activeScopes)
-                    self.start()
-                }
-            }
-        }
-        observe()
-    }
-}
-
 // MARK: - RunnerStore
 
 /// Swift 6 actor that owns the GitHub poll loop and all derived runner/job/action state.
@@ -142,7 +74,7 @@ private final class ScopesObserver {
 /// - `preferencesStore` and `scopeStore` are `@MainActor`-isolated `Sendable` protocol
 ///   values; any read of their properties must happen inside `await MainActor.run { }`.
 /// - After every fetch cycle, results are pushed to the injected `RunnerViewModel` on the
-///   main actor via `await MainActor.run { }`. SwiftUI’s `@Observable` machinery
+///   main actor via `await MainActor.run { }`. SwiftUI's `@Observable` machinery
 ///   picks up the mutation automatically — no Combine `PassthroughSubject` needed.
 /// - `LocalRunnerStore` is an `actor`; its state is read via the main-actor snapshot
 ///   pushed to `RunnerViewModel`, not by crossing the actor boundary synchronously.
@@ -171,7 +103,7 @@ actor RunnerStore {
     /// IDs of action groups whose failure hook has already fired.
     ///
     /// Kept separate from `actionGroupCache` so that cache eviction does not re-arm
-    /// the hook for old completed groups still present in GitHub’s last-completed feed.
+    /// the hook for old completed groups still present in GitHub's last-completed feed.
     private var seenGroupIDs: Set<String> = []
 
     /// Whether the GitHub API is currently rate-limiting this client.
@@ -210,15 +142,22 @@ actor RunnerStore {
     private let onStatusUpdate: @MainActor @Sendable () -> Void
 
     /// Shared `JSONDecoder` — reused across all decode calls in the actor.
-    /// Declared as a stored instance property so it is serialised by the actor’s own
+    /// Declared as a stored instance property so it is serialised by the actor's own
     /// executor; no concurrent access is possible. Used by `backfillSteps` in
     /// `RunnerStore+PollBridge.swift`.
     let decoder = JSONDecoder()
 
+    /// Fetcher for workflow action groups. Injected at init so the polling path
+    /// is testable without live network access.
+    ///
+    /// Production callers use the default (`sharedGitHubTransport`); unit tests
+    /// substitute a `StubTransport` or other `GitHubTransportProtocol` conformer.
+    let actionGroupFetcher: WorkflowActionGroupFetcher
+
     // MARK: - Aggregate status
 
     /// The combined health status across all runners, derived from the current `runners` array.
-    /// Read by external consumers (e.g. `AppDelegate`) outside this file’s analysis scope.
+    /// Read by external consumers (e.g. `AppDelegate`) outside this file's analysis scope.
     /// periphery:ignore
     var aggregateStatus: AggregateStatus { AggregateStatus(runners: runners) }
 
@@ -247,7 +186,7 @@ actor RunnerStore {
     /// - Note: Swift 6 actor `init` is nonisolated. The observation task `Task` handles
     ///   are assigned synchronously inside `startObservingPreferences` /
     ///   `startObservingScopes` (in the calling function body, before any suspension in
-    ///   the task’s async work). This means `deinit` always holds real `Task` values by
+    ///   the task's async work). This means `deinit` always holds real `Task` values by
     ///   the time any suspension occurs, even if the actor is deallocated immediately after
     ///   `init`.
     init(
@@ -255,13 +194,15 @@ actor RunnerStore {
         localRunnerStore: LocalRunnerStore,
         preferencesStore: any AppPreferencesStoreProtocol,
         scopeStore: any ScopeStoreProtocol,
-        onStatusUpdate: @escaping @MainActor @Sendable () -> Void
+        onStatusUpdate: @escaping @MainActor @Sendable () -> Void,
+        actionGroupFetcher: WorkflowActionGroupFetcher = WorkflowActionGroupFetcher()
     ) {
         self.viewModel = viewModel
         self.localRunnerStore = localRunnerStore
         self.preferencesStore = preferencesStore
         self.scopeStore = scopeStore
         self.onStatusUpdate = onStatusUpdate
+        self.actionGroupFetcher = actionGroupFetcher
         Task { await self.startObservingPreferences() }
         Task { await self.startObservingScopes() }
     }
@@ -277,14 +218,14 @@ actor RunnerStore {
     /// Starts (or restarts) the `pollingInterval` observation loop.
     ///
     /// `pollLoop.intervalObservationTask` is assigned synchronously in this function body —
-    /// before the task’s async work runs — so `deinit` always cancels a real `Task`
+    /// before the task's async work runs — so `deinit` always cancels a real `Task`
     /// rather than a `nil` optional, even if the actor is deallocated immediately
     /// after `init`.
     ///
     /// `AsyncStream.makeStream` returns the stream and a separate continuation value.
     /// The continuation is handed to `PreferencesObserver`, a `@MainActor` class that
     /// owns the recursive `withObservationTracking` registration entirely on the main
-    /// actor. The observer is returned from `MainActor.run` and held in the Task’s
+    /// actor. The observer is returned from `MainActor.run` and held in the Task's
     /// async scope so it stays alive for the full lifetime of the stream — without
     /// this, `[weak self]` in `onChange` would find `self == nil` on the first change
     /// and silently stop all future polling-interval updates.
@@ -317,13 +258,13 @@ actor RunnerStore {
     /// Starts (or restarts) the `activeScopes` observation loop.
     ///
     /// `pollLoop.scopeObservationTask` is assigned synchronously in this function body —
-    /// before the task’s async work runs — so `deinit` always cancels a real `Task`
+    /// before the task's async work runs — so `deinit` always cancels a real `Task`
     /// rather than a `nil` optional, even if the actor is deallocated immediately
     /// after `init`.
     ///
-    /// Same approach as `startObservingPreferences` — see that method’s
+    /// Same approach as `startObservingPreferences` — see that method's
     /// doc-comment for the full rationale, including why the observer must be
-    /// retained in the Task’s async scope beyond the `MainActor.run` closure.
+    /// retained in the Task's async scope beyond the `MainActor.run` closure.
     private func startObservingScopes() {
         let injectedStore = scopeStore
         pollLoop.setScopeObservationTask(Task { [weak self] in
@@ -396,7 +337,7 @@ actor RunnerStore {
     }
 
     /// Computes the delay before the next poll: 10 s while jobs/actions are active,
-    /// otherwise the user’s configured idle interval (clamped to ≥ 10 s). Also widened
+    /// otherwise the user's configured idle interval (clamped to ≥ 10 s). Also widened
     /// to the idle interval while rate-limited.
     ///
     /// `async` because it reads `preferencesStore.pollingInterval` which is
